pyproject.toml
[project]
name = "hh-bot"
version = "0.6.0"
description = "llmBot4HH"
requires-python = ">=3.11"
dependencies = [
  "playwright>=1.54.0",
  "PyYAML>=6.0.2",
  "httpx>=0.27.0",
]

[project.scripts]
bot = "bot.cli.main:main"
